+++ services/federator/src/Federator/Remote.hs
@@ -24,23 +24,24 @@ import Data.String.Conversions (cs)
 import Federator.Discovery (DiscoverFederator, LookupError, discoverFederator)
 import Imports
 import Mu.GRpc.Client.Optics (GRpcReply)
-import Mu.GRpc.Client.Record (GRpcMessageProtocol (MsgProtoBuf), GrpcClient, grpcClientConfigSimple, setupGrpcClient')
+import Mu.GRpc.Client.Record (GRpcMessageProtocol (MsgProtoBuf))
 import Mu.GRpc.Client.TyApps (gRpcCall)
-import Network.HTTP2.Client (ClientError)
+import Network.GRPC.Client.Helpers
 import Polysemy
 import Polysemy.TinyLog (TinyLog)
 import qualified Polysemy.TinyLog as Log
 import qualified System.Logger.Message as Log
+import Wire.API.Federation.GRPC.Client
 import Wire.API.Federation.GRPC.Types
 import Wire.Network.DNS.SRV (SrvTarget (SrvTarget))
 
 data RemoteError
   = RemoteErrorDiscoveryFailure LookupError Domain
-  | RemoteErrorClientFailure ClientError SrvTarget
+  | RemoteErrorClientFailure GrpcClientErr SrvTarget
   deriving (Show, Eq)
 
 data Remote m a where
-  DiscoverAndCall :: ValidatedFederatedRequest -> Remote m (Either RemoteError (GRpcReply Response))
+  DiscoverAndCall :: ValidatedFederatedRequest -> Remote m (Either RemoteError (GRpcReply InwardResponse))
 
 makeSem ''Remote
 
@@ -56,22 +57,28 @@ interpretRemote = interpret $ \case
             . Log.field "error" (show err)
         pure $ Left (RemoteErrorDiscoveryFailure err vDomain)
       Right target -> do
-        -- FUTUREWORK(federation): Make this use TLS, maybe make it configurable
-        -- FUTUREWORK(federation): Cache this client and use it for many requests
         eitherClient <- mkGrpcClient target
         case eitherClient of
           Right client ->
             Right <$> callInward client vRequest
           Left err -> pure $ Left err
 
-callInward :: MonadIO m => GrpcClient -> Request -> m (GRpcReply Response)
+callInward :: MonadIO m => GrpcClient -> Request -> m (GRpcReply InwardResponse)
 callInward client request =
   liftIO $ gRpcCall @'MsgProtoBuf @Inward @"Inward" @"call" client request
 
+-- FUTUREWORK(federation): Make this use TLS with real certificate validation
+-- FUTUREWORK(federation): Allow a configurable trust store to be used in TLS certificate validation
+--   See also https://github.com/lucasdicioccio/http2-client/issues/76
+-- FUTUREWORK(federation): Cache this client and use it for many requests
 mkGrpcClient :: Members '[Embed IO, TinyLog] r => SrvTarget -> Sem r (Either RemoteError GrpcClient)
 mkGrpcClient target@(SrvTarget host port) = do
-  let cfg = grpcClientConfigSimple (cs host) (fromInteger $ toInteger port) False
-  eitherClient <- setupGrpcClient' cfg
+  -- FUTUREWORK(federation): grpcClientConfigSimple using TLS is INSECURE and IGNORES any certificates and there's no way
+  -- to change that (at least not when using the default functions from mu or http2-grpc-client)
+  -- See https://github.com/haskell-grpc-native/http2-grpc-haskell/issues/47
+  -- While early testing, this is "convenient" but needs to be fixed!
+  let cfg = grpcClientConfigSimple (cs host) (fromInteger $ toInteger port) True
+  eitherClient <- createGrpcClient cfg
   case eitherClient of
     Left err -> do
       Log.debug $